requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
SQLAlchemy==2.0.31
WTForms==3.1.2
Werkzeug==3.0.3
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
greenlet==3.0.3
itsdangerous==2.2.0
pip==24.1
psycopg2==2.9.9
python-dotenv==1.0.1
typing_extensions==4.12.2
